SOLR-16427: Evaluate and fix errorprone rules - IntLongMath

+++ solr/core/src/java/org/apache/solr/search/SortedIntDocSet.java
@@ -750,7 +750,7 @@ public class SortedIntDocSet extends DocSet {
 
       @Override
       public long cost() {
-        return limitIdx - startIdx;
+        return (long) limitIdx - startIdx;
       }
     };
   }
